% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\title{Simulation of SIR models using R}
\usage{
sir(i0, r0, beta, gam, t_range = seq(0, 100, 1))
}
\arguments{
\item{i0}{The proportion of the population initially infected.}

\item{r0}{The proportion of the population initially recovering or dying.}

\item{beta}{The average daily effective contact of infected persons.}

\item{gam}{The average rate at which an infected person recovers (or dies).}

\item{t_range}{a vector:Time step.If null,the default is t_range=seq(1,100,1).}
}
\value{
a list including 'infe':Percentage of people infected;susc':Percentage of susceptible persons; 'reco':Percentage of people who recover (or die);t.peak':The peak of the number of infected people; 't.stop':The fastest time when the number of infected people is close to zero.
}
\description{
In the SIR model, an infected person may recover or die, and once recovered or died he or she will not be infected again.After the initial value of SIR model is given, the figure of the model is drawn and the simulation results are given.
}
\examples{
\dontrun{
SIR <- sir(i0=0.01,r0=0,beta=1.2,gam=0.4,t_range=seq(0,100,1))
plot(SIR$infe,type='l',col='red',xlab = 'time' , ylab='percentage',
main='SIR model',ylim = c(0,1))
  lines(SIR$susc,type='l',col='blue')
  lines(SIR$reco,type='l',col='green')
  legend('right',c('infective percentage','susceptible percentage',
  'recovery percentage'),col=c('red','blue','green'),lty=c(1,1,1))
}
}
