% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis.R
\name{sis}
\alias{sis}
\title{Simulation of SIS models using R}
\usage{
sis(i0, beta, gam, t_range = seq(0, 100, 1))
}
\arguments{
\item{i0}{The proportion of the population initially infected.}

\item{beta}{The average daily effective contact of infected persons.}

\item{gam}{The average rate at which an infected person recovers (or dies).}

\item{t_range}{a vector:Time step.If null,the default is t_range=seq(1,100,1).}
}
\value{
a list including 'infe':Percentage of people infected;susc':Percentage of susceptible persons; 't.stable':The number of infected people reached a minimum period of stability; 'per.stable':The percentage of people infected at a stable level.
}
\description{
In the SIS model, an infected person may recover or die, and once recovered or died he or she will not be infected again.After the initial value of SIS model is given, the figure of the model is drawn and the simulation results are given.
}
\examples{
\dontrun{
SIS <- sis(i0=0.01,beta=1.2,gam=0.4,t_range=seq(0,100,1))
plot(SIS$infe,type='l',col='red',xlab = 'time' , ylab='percentage',
main='SIS model',ylim = c(0,1))
  lines(SIS$susc,type='l',col='blue')
  legend('right',c('infective percentage','susceptible percentage'),
  col=c('red','blue'),lty=c(1,1))
}
}
